+++ internal/services/streamanalytics/helpers_input.go
@@ -95,6 +95,51 @@ func expandStreamAnalyticsStreamInputSerialization(input []interface{}) (streama
 	return nil, fmt.Errorf("Unsupported Input Type %q", inputType)
 }
 
+func expandStreamAnalyticsStreamInputSerializationTyped(serialization []Serialization) (streamanalytics.BasicSerialization, error) {
+	v := serialization[0]
+
+	inputType := streamanalytics.Type(v.Type)
+	encoding := v.Encoding
+	fieldDelimiter := v.FieldDelimiter
+
+	switch inputType {
+	case streamanalytics.TypeAvro:
+		return streamanalytics.AvroSerialization{
+			Type:       streamanalytics.TypeAvro,
+			Properties: map[string]interface{}{},
+		}, nil
+
+	case streamanalytics.TypeCsv:
+		if encoding == "" {
+			return nil, fmt.Errorf("`encoding` must be specified when `type` is set to `Csv`")
+		}
+		if fieldDelimiter == "" {
+			return nil, fmt.Errorf("`field_delimiter` must be set when `type` is set to `Csv`")
+		}
+		return streamanalytics.CsvSerialization{
+			Type: streamanalytics.TypeCsv,
+			CsvSerializationProperties: &streamanalytics.CsvSerializationProperties{
+				Encoding:       streamanalytics.Encoding(encoding),
+				FieldDelimiter: utils.String(fieldDelimiter),
+			},
+		}, nil
+
+	case streamanalytics.TypeJSON:
+		if encoding == "" {
+			return nil, fmt.Errorf("`encoding` must be specified when `type` is set to `Json`")
+		}
+
+		return streamanalytics.JSONSerialization{
+			Type: streamanalytics.TypeJSON,
+			JSONSerializationProperties: &streamanalytics.JSONSerializationProperties{
+				Encoding: streamanalytics.Encoding(encoding),
+			},
+		}, nil
+	}
+
+	return nil, fmt.Errorf("Unsupported Input Type %q", inputType)
+}
+
 func flattenStreamAnalyticsStreamInputSerialization(input streamanalytics.BasicSerialization) []interface{} {
 	var encoding string
 	var fieldDelimiter string
@@ -130,5 +175,41 @@ func flattenStreamAnalyticsStreamInputSerialization(input streamanalytics.BasicS
 			"type":            inputType,
 			"field_delimiter": fieldDelimiter,
 		},
+	}
+}
+
+func flattenStreamAnalyticsStreamInputSerializationTyped(input streamanalytics.BasicSerialization) Serialization {
+	var encoding string
+	var fieldDelimiter string
+	var inputType string
+
+	if _, ok := input.AsAvroSerialization(); ok {
+		inputType = string(streamanalytics.TypeAvro)
+	}
+
+	if v, ok := input.AsCsvSerialization(); ok {
+		if props := v.CsvSerializationProperties; props != nil {
+			encoding = string(props.Encoding)
+
+			if props.FieldDelimiter != nil {
+				fieldDelimiter = *props.FieldDelimiter
+			}
+		}
+
+		inputType = string(streamanalytics.TypeCsv)
+	}
+
+	if v, ok := input.AsJSONSerialization(); ok {
+		if props := v.JSONSerializationProperties; props != nil {
+			encoding = string(props.Encoding)
+		}
+
+		inputType = string(streamanalytics.TypeJSON)
+	}
+
+	return Serialization{
+		Encoding:       encoding,
+		Type:           inputType,
+		FieldDelimiter: fieldDelimiter,
 	}
 }